+++ server/src/main/java/org/tattour/server/domain/admin/controller/AdminController.java
@@ -2,7 +2,6 @@ package org.tattour.server.domain.admin.controller;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import java.util.Objects;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -10,7 +9,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.tattour.server.domain.order.provider.impl.OrderProviderImpl;
-import org.tattour.server.domain.point.domain.PointChargeRequest;
 import org.tattour.server.domain.point.provider.impl.PointProviderImpl;
 import org.tattour.server.global.dto.ApiResponse;
 import org.tattour.server.global.dto.SuccessType;
@@ -36,8 +34,8 @@ public class AdminController {
     @GetMapping("/point/requests")
     public ResponseEntity<?> getPointChargeRequest(
             @RequestParam(required = false) Integer userId,
-            @RequestParam(required = false) Boolean isComplete
+            @RequestParam(required = false) Boolean isCompleted
     ){
-        return ApiResponse.success(SuccessType.GET_SUCCESS, pointProvider.getAllPointChargeRequest(userId, isComplete));
+        return ApiResponse.success(SuccessType.GET_SUCCESS, pointProvider.getAllPointChargeRequest(userId, isCompleted));
     }
 }
