+++ server/src/main/java/org/tattour/server/domain/point/domain/PointChargeRequest.java
@@ -9,12 +9,14 @@ import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import lombok.AccessLevel;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.DynamicInsert;
 import org.tattour.server.domain.user.domain.User;
 
 @Entity
 @DynamicInsert
+@Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class PointChargeRequest {
     @Id
