+++ server/src/main/java/org/tattour/server/domain/point/provider/dto/response/GetPointChargeRequestListRes.java
@@ -3,7 +3,9 @@ package org.tattour.server.domain.point.provider.dto.response;
 import java.util.List;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
+import lombok.Getter;
 
+@Getter
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 public class GetPointChargeRequestListRes {
     List<GetPointChargeRequestRes> getPointChargeRequestResList;
