+++ server/src/main/java/org/tattour/server/domain/point/provider/impl/PointProviderImpl.java
@@ -2,7 +2,7 @@ package org.tattour.server.domain.point.provider.impl;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import org.tattour.server.domain.point.Dao.PointDao;
+import org.tattour.server.domain.point.dao.PointDao;
 import org.tattour.server.domain.point.provider.PointProvider;
 import org.tattour.server.domain.point.provider.dto.response.GetPointChargeRequestListRes;
 
@@ -11,8 +11,7 @@ import org.tattour.server.domain.point.provider.dto.response.GetPointChargeReque
 public class PointProviderImpl implements PointProvider {
     private final PointDao pointDao;
     @Override
-    public GetPointChargeRequestListRes getAllPointChargeRequest(Integer userId, Boolean isComplete) {
-
-        return GetPointChargeRequestListRes.of(pointDao.getPointChargeRequestResList(userId, isComplete));
+    public GetPointChargeRequestListRes getAllPointChargeRequest(Integer userId, Boolean isCompleted) {
+        return GetPointChargeRequestListRes.of(pointDao.getPointChargeRequestResList(userId, isCompleted));
     }
 }
