+++ server/src/main/java/org/tattour/server/global/util/EntityDtoMapper.java
@@ -12,7 +12,6 @@ import org.tattour.server.domain.point.domain.PointChargeRequest;
 import org.tattour.server.domain.point.provider.dto.response.GetPointChargeRequestRes;
 import org.tattour.server.domain.sticker.domain.Sticker;
 import org.tattour.server.domain.sticker.provider.dto.response.StickerLikedInfo;
-import org.tattour.server.domain.user.domain.ProductLiked;
 import org.tattour.server.domain.user.domain.User;
 import org.tattour.server.domain.user.service.dto.response.GetUserProfileRes;
 
@@ -24,6 +23,13 @@ public interface EntityDtoMapper {
     GetUserProfileRes toGetUserProfileRes(User user);
 
     // StickerLikedInfo
+    @Mapping(target = "discountPrice",
+            expression = "java(sticker.getDiscount() != null ? "
+                    + "sticker.getPrice() * (100 - sticker.getDiscount().getDiscountRate()) / 100 "
+                    + ": null)")
+    @Mapping(target = "discountRate",
+            expression = "java(sticker.getDiscount() != null ? "
+                    + "sticker.getDiscount().getDiscountRate() : null)")
     StickerLikedInfo toStickerLikedInfo(Sticker sticker);
     List<StickerLikedInfo> toStickerLikedInfoList(List<Sticker> stickerList);
